requirements.txt
# Arch
dependency-injector==4.41.0
python-dotenv==1.0.1

# Test and statis analyzsis
pytest==8.1.1
black==24.3.0
flake8==7.0.0

# Data colelct & cloud
yfinance    # yahoo finance
boto3       # aws
SQLAlchemy  # orm
psycopg2-binary  # postgresql driver
pandas      # data processing
matplotlib  # visualize
seaborn     # visualize
scikit-learn # cluster

# Data model validation
pydantic==2.6.4